requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
schedule==1.2.1
python-dotenv==1.0.1
